+++ R/biplot_data.check_model_hedonic.R
@@ -6,7 +6,7 @@
 #' @param x Output from \code{\link{check_model.fit_model_hedonic}}
 #' 
 #' @return 
-#' The function return a CA object
+#' The function return a list with two object : CA and HCPC objects
 #'  
 #' @details 
 #' S3 method.
@@ -26,8 +26,8 @@ biplot_data.check_model_hedonic = function(x){
   if( !inherits(x, "check_model_hedonic") ) {
     stop("data must come from PPBstats::check_model() with output from PPBstats:model_hedonic.")
   }
-  out = x$hedonic$CA 
-  class(out) <- c("PPBstats", "biplot_hedonic", "CA")
+  out = list("CA" = x$hedonic$CA, "HCPC" = x$hedonic$HCPC)
+  class(out) <- c("PPBstats", "biplot_hedonic")
   return(out)
 }