+++ R/format_data_PPBstats.data_organo_napping.R
@@ -34,7 +34,7 @@ format_data_PPBstats.data_organo_napping = function(data, threshold){
   if( !is.factor(data$germplasm) ) { stop("germplasm must be a factor") }
   if( !is.factor(data$location) ) { stop("location must be a factor") }
   
-  N = format_organo(data, threshold)
+  N = format_organo(data, threshold)$data_sample
   N = N[,c(6, 1, 2, 3, c(7:ncol(N)))]
   descriptors = colnames(N)[c(7:ncol(N))]