+++ R/plot.biplot_hedonic.R
@@ -12,11 +12,14 @@
 #' The plot are done with the factoextra package
 #' 
 #' @return 
-#' It returns a list with 3 elements,  which are outputs from CA analysis:
+#' It returns a list with two elements:
 #'   \itemize{
-#'    \item ind
-#'    \item var
-#'    \item axes
+#'    \item ca_biplot biplot regarding CA analysis
+#'    \item hcpc biplot biplot regarding PCA and HCPC analysis which is a list of two elements
+#'    \itemize{
+#'     \item variable of the PCA and supplementary variables
+#'     \item clusters of juges plot on the PCA
+#'    }
 #'   }
 #'   
 #' @author Pierre Riviere
@@ -32,11 +35,23 @@
 #' 
 plot.biplot_hedonic = function(x, ...){
   # see http://www.sthda.com/english/rpkgs/factoextra/reference/fviz_ca.html
+  # CA ----------
+  out_CA = x$CA
+  p = fviz_ca_biplot(out_CA)
+  p$data$sample = out_CA$call$Xtot$sample
+  ca_biplot = p + geom_point(aes(color = sample))
   
-  p = fviz_ca_biplot(x)
-  p$data$sample = x$call$Xtot$sample
-  p = p + geom_point(aes(color = sample))
+  # PCA ----------
+  out_HCPC = x$HCPC
   
-  out = list("ca_biplot" = fviz_ca_biplot(x))
+  p_var = fviz_pca_var(out_HCPC$res.pca, repel = TRUE)
+  p_var = fviz_add(p_var, out_HCPC$res.pca$quali.sup$coord, color = "red")
+  
+  hcpc_biplot = list(
+    "var" = p_var,
+    "cluster" = fviz_cluster(out_HCPC$res.hcpc, repel = TRUE) 
+  )
+  
+  out = list("ca_biplot" = ca_biplot, "hcpc_biplot" = hcpc_biplot)
   return(out)
 }